wire-protocol: simplify and fix fee updates.
There's no reason to offer a range, simply set the fee you require for
your own commitment transaction (the other node's commitment transaction is
their own problem).

Given the protocol is async, there's no way to ensure that you'll be able
to cover the fees.  Fairness is important, but robustness even more important,
so the other side must cover fees if it can.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>