wire-protocol: use explicit 64-bit HTLC identifiers, as per lnd.
(They actually insist on incrementing ids where one side is odd, the other even;
this is not that strict).

While this requires a separate lookup, it's less fragile than the
previous "order-I-added-them-in" or "by rhash" values.  "By rhash"
is definitely wrong, since we need to allow multiple HTLCs with
the same rhash, otherwise a malicious third party can probe to
discover if we're servicing a given HTLC.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>