wire-protocol: reduce to 300 HTLCs per side so we can steal in one standard tx.
Previously, I'd assumed we'd need multiple steal transactions for large
commit transactions, unless we made the HTLC limit unreasonably low.

However, since we use segregated witness and most of the giant script is in the
witness, we could have about 699 of them with a 400k cost.  So 300 each
leaves plenty of room, and seems sufficient number of HTLCs in flight for
now.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>