wire-protocol: spell out update mechanism in full detail, remove acks.
Every `update_revocation` acks the last un-revoked `update_commit`, so
we can use this to synchronize state.  The actual mechanism is slightly complex:
we have newly proposed changes (aka "unacked changes") and changes we
are applying because the remote side has sent `update_revocation` (aka
"acked changes").  That's on top the the need to track both our own state
and the state of the remote side.

The good news is that this obviates the need to have explicit acks; it's
also clear that there's nothing we can do (or even detect) if the other side
chooses to update again without waiting for our revoke.

This protocol has been validated using a simulator in the c-lightning
github tree:

https://github.com/ElementsProject/lightning/tree/test-protocol

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>